[Serialization] Extract protobuf checks into their own file
Tp keep FirSerializationPluginClassChecker small and manageable.

^KT-81042
